Extract collection request into reusable hook (#3746)
